requirements.txt
mock==2.0.0
pbr==5.1.3
pika==0.13.1
pymongo==3.7.2
python-dotenv==0.10.1
six==1.12.0
unittest-data-provider==1.0.1
